MemoryRead/WriteBuffer: Capture TaskCanceledException
There are checks that the CancellationTokenSource is cancelled, in which the result is `false`. But there was a small race condition where if before the task is properly started, and the token is cancelled, one could get a TaskCanceledException instead of false.

Need to wrap the try/catch around the `await`, as the exception appears to be raised only when the task is run, not in the context of the caller.

To reproduce, modified the test case `WaitForWriteAsyncFullCancel` to call `cts.Cancel()` (we can get rid of the task), remove the checks if it's cancelled, and the test case now passes (where it would fail).

Issue: DOTNET-439