CircularBuffer / MemoryRead/WriteBuffer: Support the Span<byte> type
Add support for the Span<byte> type for .NET Core 2.1 and later. This allows stream implementations to use the Span type to copy data.

Implementations are still a copy operation, so these new methods don't bring on their own a performance increase. They are more of a convenience so that the user doesn't need to wrap their array types around a Span<byte>.

Issue: DOTNET-341